refactor: move HttpBody to cli/http_util.rs (#3569)
